Add Package and Package version models
We want to track the version of the Stack used to generate each job.
EUPS makes this a little tricky because there's no portable version
database that can be queried. These tables attempt to capture
information sufficient to understand exactly what version of the code
was used to create a result.

The Package model has records for every EUPS package in a stack. The
Package version has records for the version of a package seen in a job.
Thus if there are N packages, there will be N rows in the PackageVersion
table for every build.

Note that we may eventually want to make use of the Stack's providence
tracking code.