Add VersionedPackage model to track Git versions
We want to track the version of the Stack used to generate each job.
EUPS makes this a little tricky because there's no portable version
database that can be queried.

The VersionedPackage table attempts to capture what versions of Stack
software were used during a QA test run. Each Job has many
VersionedPackages that it references (one-to-many).

VersionedPackages are serialized within a Job resources under the
'packages' field.

Note that originally I attempted to re-factor the VersionedPackage table
into both a Package table and a Package Version table.  The Package
model has records for every EUPS package in a stack. The PackageVersion
has records for the version of a package seen in a job.  Thus if there
are N packages, there will be N rows in the PackageVersion table for
every build.  This was implemented as a many-to-many relationship
between Jobs and Packages with the PackageVersion table as an
intermediary.  Unfortunately it was difficult (impossible?) to make this
work with the Django REST Framework as a nested serialization.